refactor: split novel workflow service facade (#11)
